refactor(playwright): rewrite node tests as user journey integration tests
Replace 16 isolated unit-style test files with 7 journey files where
every test is a multi-step user flow:

- server-bootstrap: SDK init → FHE infra check → metadata → shutdown
- shield-and-query: shield USDT/USDC → verify ERC-20 + handles
- session-lifecycle: allow → shield → revoke → re-allow → disconnect
- delegation-lifecycle: delegate → verify → revoke → edge cases
- multi-token-portfolio: shield both → metadata → wrappers → approve
- worker-pool-concurrency: parallel keypairs/EIP-712 → shutdown
- chain-switching: pool re-init on chain change → ACL switch

22 integration tests, all passing in ~14s.
