--- packages/playwright/tests/node/approve.node.spec.ts
@@ -1,16 +1,0 @@
-import { nodeTest as test, expect } from "../../fixtures/node-test";
-import type { Address } from "viem";
-
-test.describe("Token.approveUnderlying — ERC-20 approval for shield", () => {
-  test("approves wrapper to spend underlying tokens", async ({ sdk, contracts }) => {
-    const token = sdk.createToken(contracts.USDC, contracts.cUSDC as Address);
-    const result = await token.approveUnderlying(1000n * 10n ** 6n);
-    expect(result.txHash).toMatch(/^0x[0-9a-fA-F]{64}$/);
-  });
-
-  test("approves max uint256 by default", async ({ sdk, contracts }) => {
-    const token = sdk.createToken(contracts.USDT, contracts.cUSDT as Address);
-    const result = await token.approveUnderlying();
-    expect(result.txHash).toMatch(/^0x[0-9a-fA-F]{64}$/);
-  });
-});
